Fine-grained right management (in the GUI for now)

+++ src/repo-users.php
@@ -7,9 +7,11 @@ if (empty($_GET['repo'])) {
   $repo = $_GET['repo'];
 }
 $errorMsg = '';
-if ($admin && isset($_POST['submit_user_add'])) {
+$repoadmin = $admin || isrepoadmin($repo, $_SESSION['username']);
+if ($repoadmin && isset($_POST['submit_user_add'])) {
   $fUser = $_POST['username'];
-  $res = gitrepoinfo('add-user', $repo, $fUser);
+  $fRight = $_POST['right'];
+  $res = gitrepoinfo('add-user', $repo, $fUser, $fRight);
   if ($res === false) {
     $errorMsg = "L'utilisateur n'a pas pu être ajouté.";
   }
@@ -22,21 +24,30 @@ require('header.inc.php');
       <table>
         <tr>
           <th class="name">Utilisateur</th>
+          <th class="right">Droit</th>
           <th class="actions">Actions</th>
         </tr>
 <?php
-$users = array_map(function($val) { return explode(':', $val)[0]; }, gitrepoinfo('show-users', $repo));
-foreach ($users as $user) {
+$members = array();
+foreach (gitrepoinfo('show-users', $repo) as $userinfo) {
+  $info = explode(':', $userinfo);
+  $members[$info[0]] = $info[1];
+}
+$isExport = file_exists("$gitdir/$repo.git/git-daemon-export-ok");
+foreach ($members as $user => $right) {
   $actions = ' — ';
-  if ($admin) {
-    $actions = "<a href=\"/{$gitwebroot}remove_user/$repo/$user\">Retirer</a>";
+  if ($repoadmin) {
+    $actions = "<a href=\"/{$gitwebroot}user_right/$repo/$user/admin\">→ admin right</a>";
+    $actions .= "&nbsp;<a href=\"/{$gitwebroot}user_right/$repo/$user/user\">→ user right</a>";
+    $actions .= "&nbsp;<a href=\"/{$gitwebroot}user_right/$repo/$user/readonly\">→ readonly right</a>";
+    $actions .= "&nbsp;<a href=\"/{$gitwebroot}remove_user/$repo/$user\">Retirer</a>";
   }
-  echo "        <tr><td class=\"name\">$user</td><td class=\"actions\">$actions</td></tr>\n";
+  echo "        <tr><td class=\"name\">$user</td><td class=\"right\">$right</td><td class=\"actions\">$actions</td></tr>\n";
 }
 ?>
       </table>
     </div>
-<?php if ($admin) { ?>
+<?php if ($repoadmin) { ?>
     <div class="error"><?php echo $errorMsg; ?></div>
     <form id="repo-add-user" action="" method="POST">
       <fieldset>
@@ -46,11 +57,14 @@ foreach ($users as $user) {
 <?php
 $users = gitrepoinfo('list-users');
 foreach ($users as $user) {
-  $user = htmlspecialchars($user);
-  echo "          <option value=\"$user\">$user</option>\n";
+  if (!array_key_exists($user, $members)) {
+    $user = htmlspecialchars($user);
+    echo "          <option value=\"$user\">$user</option>\n";
+  }
 }
 ?>
         </select>
+        <select name="right"><option value="admin">admin</option><option value="user" selected="selected">user</option><option value="readonly">readonly</option></select>
         <input type="submit" name="submit_user_add" value="Ajouter l'utilisateur au dépôt"/>
       </fieldset>
     </form>
